pyproject.toml
[tool.poetry]
name = "tigrisdb"
version = "0.1.0"
description = "Python SDK for Tigris <http://www.tigrisdata.com>"
authors = ["Tigris team <support@tigrisdata.com>"]
repository = "https://www.github.com/tigrisdata/tigris-client-python"
documentation = "https://www.tigrisdata.com/docs/"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Topic :: Database"
]
license = "Apache-2.0"
readme = "README.md"
keywords = ["database", "nosql", "vector", "search"]
include = ["api/generated/**/*"]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
protobuf = "^4.22.3"
grpcio-tools = "^1.54.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.0"
coverage = {version = "^7.2.5", extras = ["toml"]}

[tool.poetry.scripts]
make = "scripts.proto:main"

[tool.coverage.run]
source = ["tigrisdb"]

[tool.coverage.report]
fail_under = 35

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
